5_Telegram-bot/requirements.txt
python-telegram-bot
python-dotenv
markdown2